Move user notification settings into the /users router
