+++ fixbackend/auth/schemas.py
@@ -15,6 +15,9 @@
 import uuid
 from pydantic import BaseModel, Field
 from fastapi_users import schemas
+from fixbackend.ids import UserId
+
+from fixbackend.notification.user_notification_repo import UserNotificationSettings
 
 
 class UserRead(schemas.BaseUser[uuid.UUID]):
@@ -40,3 +43,22 @@ class OAuthProviderAssociateUrl(BaseModel):
     account_id: uuid.UUID | None = Field(description="ID of the OAuth account, if associated")
     account_email: str | None = Field(description="Email of the user if already associated")
     authUrl: str = Field(description="URL to initiate association flow")
+
+
+class UserNotificationSettingsRead(BaseModel):
+    weekly_report: bool = Field(description="Whether to send a weekly report")
+    inactivity_reminder: bool = Field(description="Whether to send a reminder for open incidents")
+
+    @staticmethod
+    def from_model(model: UserNotificationSettings) -> "UserNotificationSettingsRead":
+        return UserNotificationSettingsRead(
+            weekly_report=model.weekly_report,
+            inactivity_reminder=model.inactivity_reminder,
+        )
+
+    def to_model(self, user_id: UserId) -> UserNotificationSettings:
+        return UserNotificationSettings(
+            user_id=user_id,
+            weekly_report=self.weekly_report,
+            inactivity_reminder=self.inactivity_reminder,
+        )
