+++ fixbackend/notification/notification_router.py
@@ -26,8 +26,6 @@ from fixbackend.ids import WorkspaceId, BenchmarkName
 from fixbackend.logging_context import set_workspace_id, set_context
 from fixbackend.notification.model import WorkspaceAlert, AlertingSetting, NotificationProvider
 from fixbackend.notification.notification_service import NotificationService
-from fixbackend.notification.schemas import UserNotificationSettingsRead
-from fixbackend.notification.user_notification_repo import UserNotificationSettingsReporitoryDependency
 
 log = logging.getLogger(__name__)
 AddSlack = "notification_add_slack"
@@ -251,27 +249,6 @@ def notification_router(fix: FixDependencies) -> APIRouter:
             return Response(status_code=204)
         except ValueError as ex:
             return JSONResponse(status_code=422, content=dict(error=str(ex)))
-
-    @router.get("/{workspace_id}/notification/user")
-    async def get_user_notification_settings(
-        user: AuthenticatedUser,
-        workspace_id: WorkspaceId,
-        user_notification_repo: UserNotificationSettingsReporitoryDependency,
-    ) -> UserNotificationSettingsRead:
-        settings = await user_notification_repo.get_notification_settings(user.id)
-        return UserNotificationSettingsRead.from_model(settings)
-
-    @router.put("/{workspace_id}/notification/user")
-    async def update_user_notification_settings(
-        user: AuthenticatedUser,
-        workspace_id: WorkspaceId,
-        notification_settings: UserNotificationSettingsRead,
-        user_notification_repo: UserNotificationSettingsReporitoryDependency,
-    ) -> UserNotificationSettingsRead:
-        updated = await user_notification_repo.update_notification_settings(
-            user.id, notification_settings.to_model(user.id)
-        )
-        return UserNotificationSettingsRead.from_model(updated)
 
     async def _delete_notification_provider_config(
         workspace_id: WorkspaceId, provider: NotificationProvider
